# Conflicts:
#	src/SFA.DAS.CandidateAccount.Data/WorkExperience/WorkHistoryRepository.cs

+++ src/SFA.DAS.CandidateAccount.Domain/Application/WorkHistoryEntity.cs
@@ -3,7 +3,7 @@
     public class WorkHistoryEntity
     {
         public Guid Id { get; set; }
-        public short WorkHistoryType { get; set; }
+        public byte WorkHistoryType { get; set; }
         public string Employer { get; set; }
         public string JobTitle { get; set; }
         public DateTime StartDate { get; set; }
